Cleans Code / Adds PID Tuning on-the-fly
After last night (2/19/22) of PID tuning, we realized this
needs to be way faster, most notably not waiting for the robot
code to deploy after every single change, so this will allow us
to change some Constants (PID and Min) from the dashboard.

Also removes some testing code

And removes extraneous field in the AutoAlign Commands.